+++ app/src/main/java/space/taran/arkretouch/filter/FilterItem.kt
@@ -1,4 +1,4 @@
-package space.taran.arkretouch
+package space.taran.arkretouch.filter
 
 import android.graphics.Bitmap
 import com.zomato.photofilters.imageprocessors.Filter
