+++ app/src/main/res/layout/activity_edit.xml
@@ -22,7 +22,7 @@
         app:cropBackgroundColor="@color/crop_image_view_background"
         app:cropInitialCropWindowPaddingRatio="0"/>
 
-    <space.taran.arkretouch.EditorDrawCanvas
+    <space.taran.arkretouch.paint.EditorDrawCanvas
         android:id="@+id/editor_draw_canvas"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
